feat: support unique assertion types (#350)
This adds support for the soon to be release client-authentication+jwt JSON Web Token Header Type and accompanying application/client-authentication+jwt MIME type. This type is intentional on exclusively being used for client assertions. In addition for the time being it allows client assertions to use a header absent the type.

Fixes #348