feat: redirect uri matching strategy (#377)
This adds support for per-client redirect URI matching strategies. The default and recommended is the fallback for all clients which uses simple string comparison. A origin based strategy exists that will determine if the requested 'redirect_uri' value shares the same origin as the registered one, as long as the registered 'redirect_uri' is only an origin without any additional URI parts.